Code review fixups
