+++ packages/server/dataloader/primaryKeyLoaderMakers.ts
@@ -14,6 +14,7 @@ import {
   selectNewFeatures,
   selectNewMeetings,
   selectNotifications,
+  selectOAuthAPIProvider,
   selectOrganizations,
   selectPages,
   selectReflectPrompts,
@@ -187,4 +188,8 @@ export const teamNotificationSettings = primaryKeyLoaderMaker((ids: readonly num
 
 export const pages = primaryKeyLoaderMaker((ids: readonly number[]) => {
   return selectPages().where('id', 'in', ids).execute()
+})
+
+export const oAuthProviders = primaryKeyLoaderMaker((ids: readonly string[]) => {
+  return selectOAuthAPIProvider().where('id', 'in', ids).execute()
 })
