+++ packages/server/graphql/public/permissions.ts
@@ -38,6 +38,10 @@ const permissionMap: PermissionMap<Resolvers> = {
     addOrg: rateLimit({perMinute: 2, perHour: 5}),
     addTeam: rateLimit({perMinute: 15, perHour: 50}),
     createImposterToken: isSuperUser,
+    createOAuthAPIProvider: hasOrgRole<'Mutation.createOAuthAPIProvider'>(
+      'args.input.orgId',
+      'ORG_ADMIN'
+    ),
     loginWithGoogle: and(
       not(isEnvVarTrue('AUTH_GOOGLE_DISABLED')),
       rateLimit({perMinute: 50, perHour: 500})
