+++ packages/server/graphql/public/types/Organization.ts
@@ -1,3 +1,4 @@
+import {selectOAuthAPIProvider} from '../../../postgres/select'
 import {
   getUserId,
   isSuperUser,
@@ -120,9 +121,10 @@ const Organization: OrganizationResolvers = {
   orgFeatureFlags: async ({id: orgId}, _args, {dataLoader}) => {
     return dataLoader.get('allFeatureFlagsByOwner').load({ownerId: orgId, scope: 'Organization'})
   },
-  oauthProviders: async ({id: orgId}) => {
-    const {default: oauthProviders} = await import('./Organization/oauthProviders')
-    return oauthProviders({id: orgId})
+  oauthProviders: async ({id: orgId}, _args, {dataLoader}) => {
+    return dataLoader
+      ? await dataLoader.get('oauthProvidersByOrgId').load(orgId)
+      : await selectOAuthAPIProvider().where('orgId', '=', orgId).execute()
   }
 }
