+++ packages/server/server.ts
@@ -14,8 +14,8 @@ import mattermostWebhookHandler from './integrations/mattermost/mattermostWebhoo
 import jiraImagesHandler from './jiraImagesHandler'
 import listenHandler from './listenHandler'
 import {metricsHandler} from './metricsHandler'
-import authorizeHandler from './oauth/authorizeHandler'
-import tokenHandler from './oauth/tokenHandler'
+import authorizeHandler from './oauth2/authorizeHandler'
+import tokenHandler from './oauth2/tokenHandler'
 import selfHostedHandler from './selfHostedHandler'
 import {createStaticFileHandler} from './staticFileHandler'
 import getReqAuth from './utils/getReqAuth'
